+++ contracts/canonical/CrunaRegistry.sol
@@ -11,12 +11,17 @@ pragma solidity ^0.8.20;
 
 import {ICrunaRegistry} from "./ICrunaRegistry.sol";
 
+/**
+  @title CrunaRegistry
+  @dev Manages the creation of token bound accounts
+*/
 contract CrunaRegistry is ICrunaRegistry {
   /**
-   * @dev The registry MUST revert with TokenLinkedContractCreationFailed error if the create2 operation fails.
+     @dev The registry MUST revert with TokenLinkedContractCreationFailed error if the create2 operation fails.
    */
   error TokenLinkedContractCreationFailed();
 
+  /// @dev see {ICrunaRegistry-createTokenLinkedContract}
   function createTokenLinkedContract(
     address implementation,
     bytes32 salt,
@@ -93,6 +98,7 @@ contract CrunaRegistry is ICrunaRegistry {
     }
   }
 
+  /// @dev see {ICrunaRegistry-tokenLinkedContract}
   function tokenLinkedContract(
     address implementation,
     bytes32 salt,