+++ contracts/manager/Actor.sol
@@ -5,20 +5,44 @@ pragma solidity ^0.8.20;
 
 // import {console} from "hardhat/console.sol";
 
-// @dev This contract manages actors
+/**
+  @title Actor
+  @dev This contract manages actors (protectors, safe recipients, sentinels, etc.)
+*/
 contract Actor {
+  /// @dev The maximum number of actors that can be set
   uint256 private constant _MAX_ACTORS = 16;
+
+  /// @dev The actors for each role
   mapping(bytes4 role => address[] actors) internal _actors;
 
+  /// @dev Error returned when trying to add a zero address
   error ZeroAddress();
+
+  /// @dev Error returned when trying to add an actor already added
   error ActorAlreadyAdded();
+
+  /// @dev Error returned when trying to add too many actors
   error TooManyActors();
+
+  /// @dev Error returned when an actor is not found
   error ActorNotFound();
 
+  /**
+    @dev Returns the actors for a role
+    @param role The role
+    @return The actors
+  */
   function _getActors(bytes4 role) internal view virtual returns (address[] memory) {
     return _actors[role];
   }
 
+  /**
+    @dev Returns the index of an actor for a role
+    @param actor_ The actor
+    @param role The role
+    @return The index. If the index == _MAX_ACTORS, the actor is not found
+  */
   function _actorIndex(address actor_, bytes4 role) internal view virtual returns (uint256) {
     address[] storage actors = _actors[role];
     // This may go out of gas if there are too many actors
@@ -34,20 +58,41 @@ contract Actor {
     return _MAX_ACTORS;
   }
 
+  /**
+    @dev Returns the number of actors for a role
+    @param role The role
+    @return The number of actors
+  */
   function _actorCount(bytes4 role) internal view virtual returns (uint256) {
     return _actors[role].length;
   }
 
+  /**
+    @dev Returns if an actor is active for a role
+    @param actor_ The actor
+    @param role The role
+    @return If the actor is active
+  */
   function _isActiveActor(address actor_, bytes4 role) internal view virtual returns (bool) {
     uint256 i = _actorIndex(actor_, role);
     return i < _MAX_ACTORS;
   }
 
+  /**
+    @dev Removes an actor for a role
+    @param actor_ The actor
+    @param role The role
+  */
   function _removeActor(address actor_, bytes4 role) internal virtual {
     uint256 i = _actorIndex(actor_, role);
     _removeActorByIndex(i, role);
   }
 
+  /**
+    @dev Removes an actor for a role by index
+    @param i The index
+    @param role The role
+  */
   function _removeActorByIndex(uint256 i, bytes4 role) internal virtual {
     address[] storage actors = _actors[role];
     unchecked {
@@ -59,6 +104,11 @@ contract Actor {
     actors.pop();
   }
 
+  /**
+    @dev Adds an actor for a role
+    @param actor_ The actor
+    @param role_ The role
+  */
   function _addActor(address actor_, bytes4 role_) internal virtual {
     if (actor_ == address(0)) revert ZeroAddress();
     // We allow to add up to 16 actors per role per owner to avoid the risk of going out of gas
@@ -69,6 +119,10 @@ contract Actor {
     _actors[role_].push(actor_);
   }
 
+  /**
+    @dev Deletes all the actors for a role
+    @param role The role
+  */
   function _deleteActors(bytes4 role) internal virtual {
     delete _actors[role];
   }