+++ contracts/plugins/ICrunaPlugin.sol
@@ -12,6 +12,9 @@ import {CrunaManager} from "../manager/CrunaManager.sol";
    the primary manage, which is CrunaManager.sol
 */
 interface ICrunaPlugin is ITokenLinkedContract, IVersioned {
+  /**
+    @dev The configuration of the plugin
+  */
   struct Conf {
     CrunaManager manager;
     // When mustReset is true, the plugin must be reset before being used again.
@@ -20,33 +23,60 @@ interface ICrunaPlugin is ITokenLinkedContract, IVersioned {
     uint32 mustBeReset;
   }
 
+  /**
+    @dev Error returned when the plugin is reset
+    @param implementation The address of the new implementation
+  */
   error UntrustedImplementation(address implementation);
-  error InvalidVersion(uint256 version);
+
+  /**
+    @dev Error returned when the plugin is reset
+    @param oldVersion The version of the current implementation
+    @param newVersion The version of the new implementation
+  */
+  error InvalidVersion(uint256 oldVersion, uint256 newVersion);
+
+  /**
+    @dev Error returned when the plugin is reset
+    @param requiredVersion The version required by the plugin
+  */
   error PluginRequiresUpdatedManager(uint256 requiredVersion);
+
+  /// @dev Error returned when the plugin is reset
   error Forbidden();
+
+  /// @dev Error returned when the plugin must be reset before using it
   error PluginMustBeReset();
 
+  /// @dev Initialize the plugin. It must be implemented, but can do nothing is no init is needed.
   function init() external;
 
-  // function called in the dashboard to know if the plugin is asking the
-  // right to make a managed transfer of the vault
+  /**
+    @dev Called by the manager during the plugging to know if the plugin is asking the
+      right to make a managed transfer of the vault
+  */
   function requiresToManageTransfer() external pure returns (bool);
 
+  /// @dev Called by the manager to know it the plugin must be reset when transferring the NFT
   function requiresResetOnTransfer() external pure returns (bool);
 
+  /// @dev Called by the manager to know if the plugin is an ERC721 account
   function isERC6551Account() external pure returns (bool);
 
-  // Reset the plugin to the factory settings
+  /// @dev Reset the plugin to the factory settings
   function reset() external;
 
-  // here is were _conf.mustBeReset is set to 1
+  // @dev During transfer, to reduce gas consumption, should set _conf.mustBeReset to 1
   function resetOnTransfer() external;
 
-  // @dev Upgrade the implementation of the manager/plugin
-  //   Notice that the owner can upgrade active or disable plugins
-  //   so that, if a plugin is compromised, the user can disable it,
-  //   wait for a new trusted implementation and upgrade it.
+  /**
+    @dev Upgrade the implementation of the manager/plugin
+      Notice that the owner can upgrade active or disable plugins
+      so that, if a plugin is compromised, the user can disable it,
+      wait for a new trusted implementation and upgrade it.
+  */
   function upgrade(address implementation_) external;
 
+  /// @dev Returns the manager
   function manager() external view returns (CrunaManager);
 }