+++ contracts/utils/ICommonBase.sol
@@ -1,14 +1,15 @@
 // SPDX-License-Identifier: GPL3
 pragma solidity ^0.8.20;
 
-// Author: Francesco Sullo <francesco@sullo.co>
-
 import {CrunaProtectedNFT} from "../token/CrunaProtectedNFT.sol";
 
 // import {console} from "hardhat/console.sol";
 
+/// @title ICommonBase
 interface ICommonBase {
+  /// @dev Error returned when the caller is not the token owner
   error NotTheTokenOwner();
 
+  /// @dev Returns the vault, i.e., the CrunaProtectedNFT contract
   function vault() external view returns (CrunaProtectedNFT);
 }