+++ contracts/utils/INamedAndVersioned.sol
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: GPL3
 pragma solidity ^0.8.20;
 
-// Author: Francesco Sullo <francesco@sullo.co>
-
 import {IVersioned} from "./IVersioned.sol";
 import {INamed} from "./INamed.sol";
 
+/// @title INamedAndVersioned
+/// @dev Combines INamed and IVersioned
 interface INamedAndVersioned is INamed, IVersioned {}