+++ libs-canonical/not-localhost/Canonical.sol
@@ -7,9 +7,13 @@ import {ICrunaGuardian} from "../canonical/ICrunaGuardian.sol";
 import {ICrunaRegistry} from "../canonical/CrunaRegistry.sol";
 
 /**
-  Canonical for testing testnet and mainnet
+  @title Canonical
+  @dev Returns the address where registries and guardian have been deployed
+    There are two set of addresses. In both, the registry are on the same addresses, but
+    the guardian has a different address for testing and deployment to localhost (using the 2nd
+    and 3rd hardhat standard wallets as proposer and executor).
+    This contract is for production and replaces the dev version when publishing to Npm.
 */
-
 library Canonical {
 
   function crunaRegistry() internal pure returns (ICrunaRegistry) {